changed: reading command and arguments in module 'main' added: validation and tests for module 'command_parser'

+++ handbook/command_parser.py
@@ -1,6 +1,7 @@
 from enum import Enum
 
 from handbook.customer_service import CustomerService
+from handbook.validator import Validator
 
 
 class Command(Enum):
@@ -30,30 +31,31 @@ class Command(Enum):
     )
     INSERT = (
         "insert",
-        "Error: 'insert' command requires 5 positional arguments\n"
-        "< insert customer_id full_name position name_of_the_organization email phone >"
+        "Error: 'insert' command requires 5 positional arguments separated by commas (no spaces!)\n"
+        "<insert customer_id,full_name,position,name_of_the_organization,email,phone>"
     )
     UPDATE = (
         "update",
-        "Error: 'update' command requires 5 positional arguments\n"
-        "< update customer_id full_name position name_of_the_organization email phone >"
+        "Error: 'update' command requires 5 positional arguments separated by commas (no spaces!)\n"
+        "<update customer_id,full_name,position,name_of_the_organization,email,phone>"
     )
     FIND = (
         "find",
-        "Error: 'find' command requires positional argument name (customer_id,"
-        "full_name, position name,_of_the_organization, email, phone) and argument value\n"
-        "< find 'one of the customer arguments' 'argument value'>"
+        "Error: 'find' command requires one positional argument name (customer_id,"
+        "full_name, position name,_of_the_organization, email, phone) and one argument value "
+        "separated by commas (no spaces!)\n"
+        "<find 'one of the customer arguments','argument value'>"
     )
     DELETE = (
         "delete",
         "Error: 'delete' command requires 1 positional arguments - customer_id\n"
-        "< delete customer_id >"
+        "<delete customer_id>"
     )
     LIST = (
         "list",
         "Error: 'list' command requires any number positional arguments (customer_id,"
-        "full_name, position name,_of_the_organization, email, phone)\n"
-        "< list 'any number of customer arguments separated by a space' >"
+        "full_name, position name,_of_the_organization, email, phone) separated by commas (no spaces!)\n"
+        "<list 'any number of customer arguments separated by a space'>"
     )
 
     @staticmethod
@@ -66,6 +68,7 @@ class Command(Enum):
 class Parser:
     def __init__(self):
         self.customer_service = CustomerService()
+        self.validator = Validator()
 
     def parse_command(self, command, arguments):
         if command == Command.HELP:
@@ -73,31 +76,55 @@ class Parser:
         elif command == Command.EXIT:
             raise SystemExit
         elif command == Command.INSERT:
-            try:
-                self.customer_service.insert_customer(*arguments)
-            except TypeError:
-                print(command.INSERT.description)
+            valid_arguments = dict()
+            valid_arguments['customer_id'] = arguments[0]
+            valid_arguments['full_name'] = arguments[1]
+            valid_arguments['position'] = arguments[2]
+            valid_arguments['name_of_the_organization'] = arguments[3]
+            valid_arguments['email'] = arguments[4]
+            valid_arguments['phone'] = arguments[5]
+            if self.validator.validate_data(valid_arguments):
+                try:
+                    self.customer_service.insert_customer(*arguments)
+                except TypeError:
+                    print(command.INSERT.description)
         elif command == Command.UPDATE:
-            try:
-                self.customer_service.update_customer(*arguments)
-            except TypeError:
-                print(command.UPDATE.description)
+            valid_arguments = dict()
+            valid_arguments['customer_id'] = arguments[0]
+            valid_arguments['full_name'] = arguments[1]
+            valid_arguments['position'] = arguments[2]
+            valid_arguments['name_of_the_organization'] = arguments[3]
+            valid_arguments['email'] = arguments[4]
+            valid_arguments['phone'] = arguments[5]
+            if self.validator.validate_data(valid_arguments):
+                try:
+                    self.customer_service.update_customer(*arguments)
+                except TypeError:
+                    print(command.UPDATE.description)
         elif command == Command.DELETE:
-            try:
-                self.customer_service.delete_customer(*arguments)
-            except TypeError:
-                print(command.DELETE.description)
+            valid_arguments = dict()
+            valid_arguments['customer_id'] = arguments[0]
+            if self.validator.validate_data(valid_arguments):
+                try:
+                    self.customer_service.delete_customer(*arguments)
+                except TypeError:
+                    print(command.DELETE.description)
         elif command == Command.LIST:
-            try:
-                list_of_customer = self.customer_service.list_of_customer(arguments)
-                print(*list_of_customer, sep='\n')
-            except TypeError:
-                print(command.LIST.description)
+            if self.validator.validate_data_for_list(arguments):
+                try:
+                    list_of_customer = self.customer_service.list_of_customer(arguments)
+                    print(*list_of_customer, sep='\n')
+                except TypeError:
+                    print(command.LIST.description)
         elif command == Command.FIND:
-            try:
-                customer = self.customer_service.find_customer(*arguments)
-                print(customer)
-            except TypeError:
-                print(command.FIND.description)
+            valid_arguments = dict()
+            argument_name, argument_value = arguments[0], arguments[1]
+            valid_arguments[argument_name] = argument_value
+            if self.validator.validate_data(valid_arguments):
+                try:
+                    customer = self.customer_service.find_customer(*arguments)
+                    print(customer)
+                except TypeError:
+                    print(command.FIND.description)
         else:
             print('Invalid command!')
